authentication/urls.py
from django.urls import path, include
from .views import UserViewSet
from rest_framework import routers

router = routers.DefaultRouter()
router.register('users', UserViewSet)

urlpatterns = [
    path('', include(router.urls)),
    path('auth/', include('rest_auth.urls'), name='auth'),
    path('auth/registration/', include('rest_auth.registration.urls')),
]