requirements.txt
djangorestframework==3.11.0
django-rest-auth==0.9.5
django-allauth==0.41.0
django-cors-headers==3.2.1
Django==2.2.13
psycopg2
gunicorn
dj-database-url==0.5.0
django-heroku
django-cleanup
Pillow